requirements.txt
aiogram
whisper==1.1.10
python-dotenv
googletrans==4.0.0rc1
pydub
uuid
git+https://github.com/openai/whisper.git
